Automatic dhcp_scope detection
Every network now can loosely have a 'dhcp_scope' value associated with
it. This is accomplished via first looking in the network's keyvalue
store for a 'dhcp_scope' key and if that key doesn't exist, calculate a
possible scope value by looking at the network's site/vlan.

During SREG creation the 'dhcp_scope' key will be set on any new
HWAdapters automatically. The value will be set to the dhcp_scope
value of the network of the range that is selected from the ip-chooser
widget.

DHCP scope key mismatches are highlighted in red.