Create nameserver records *before* associating SOA record
Inventory will error out if you try to create a zone (asign a new SOA to
a domain) and do not have NS records in the zone. This patch changes the
create_zone behavior to create NS records for a zone's root domain
before assinging the new SOA record. This also means dns build jobs are
not scheduled when the NS records are added (because their domain
doesn't have a SOA yet) so the tests were also updated to reflect that
change.